config: Reorganize config code and add config command
In order to simplify the config code and provide people
an easier way to edit their configs, move most of the writing
logic out of the read() fn and into write() and set_value()
functions. Running normal revup commands will no longer write
to the config, instead only "revup config" will be able
to write arbitrary config values. This means that users will
no longer be required to enter an oauth for commands that
don't talk to github, and if a command requires oauth, it
will instead prompt users to run "revup config".

Topic:config